+++ packages/lib/src/lua/mod.rs
@@ -1,2 +1,3 @@
 pub mod net;
+pub mod stdio;
 pub mod task;